// apps/finance/src/app/remote-entry/model/accountlist.interface.ts
export interface ACCOUNTLIST {
  id: number;
  description: string;
}

export interface ACCOUNTPOPUP{
  alias: string;
  name: string;
  id: number;
}

export interface ACCOUNTBYLIST{
  // listID: number;
  // accountID: number;
  // branchID: number;
  alias: string;
  name: string;
  id: number;
}